# services/indexer/openapi.yaml
openapi: 3.0.0
info:
  title: Fund My Cause Indexer REST API
  version: 1.0.0
  description: Read-only REST API for campaign data indexed from the blockchain

servers:
  - url: http://localhost:3001
  - url: https://api.fundmycause.com

paths:
  /campaigns:
    get:
      summary: List all campaigns
      tags:
        - Campaigns
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, total_raised, status]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  /campaigns/{id}:
    get:
      summary: Get campaign details
      tags:
        - Campaigns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: Campaign ID or contract address
      responses:
        '200':
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Campaign not found

  /campaigns/{id}/contributions:
    get:
      summary: List campaign contributions
      tags:
        - Contributions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of contributions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contribution'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  /stats:
    get:
      summary: Global platform statistics
      tags:
        - Analytics
      responses:
        '200':
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'

components:
  schemas:
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creator_address:
          type: string
        contract_address:
          type: string
        title:
          type: string
        description:
          type: string
        goal:
          type: integer
        total_raised:
          type: integer
        status:
          type: string
          enum: [active, succeeded, failed]
        deadline:
          type: integer
        contributor_count:
          type: integer
        created_at:
          type: string
          format: date-time

    Contribution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contributor_address:
          type: string
        amount:
          type: integer
        token_amount:
          type: integer
        contributed_at:
          type: string
          format: date-time
        tx_hash:
          type: string

    Stats:
      type: object
      properties:
        total_campaigns:
          type: integer
        active_campaigns:
          type: integer
        succeeded_campaigns:
          type: integer
        total_raised:
          type: integer
        total_contributors:
          type: integer
        avg_raised_per_campaign:
          type: number

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer